requirements.txt
pandas==2.2.3
google==3.0.0
google-api-core==2.25.1
google-auth==2.40.3
google-cloud==0.34.0
google-cloud-core==2.4.3
google-cloud-storage==3.2.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
sentence-transformers==5.0.0
faiss-cpu==1.11.0
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-text-splitters==0.3.8
langchain-huggingface==0.3.0
langchain-openai==0.3.27
python-dotenv == 1.0.1
fastapi[standard]
uvicorn == 0.34.0
openai==1.95.1
tiktoken==0.9.0